requirements.txt
requests
beautifulsoup4
pandas
numpy
openai
python-dotenv